// bank/src/types/db.d.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      states: {
        Row: {
          code: string
          name: string
        }
        Insert: {
          code: string
          name: string
        }
        Update: {
          code?: string
          name?: string
        }
        Relationships: []
      }
// ...
      users: {
        Row: {
          city: string
          created_at: string
          email: string
          first_name: string
          id: number
          last_name: string
          phone: string
          role: Database["public"]["Enums"]["user_role_enum"]
          state: string
          status: Database["public"]["Enums"]["user_status_enum"]
          street: string
          zipcode: string
        }
        Insert: {
          city: string
          created_at?: string
          email: string
          first_name: string
          id?: number
          last_name: string
          phone: string
          role?: Database["public"]["Enums"]["user_role_enum"]
          state: string
          status?: Database["public"]["Enums"]["user_status_enum"]
          street: string
          zipcode: string
        }
        Update: {
          city?: string
          created_at?: string
          email?: string
          first_name?: string
          id?: number
          last_name?: string
          phone?: string
          role?: Database["public"]["Enums"]["user_role_enum"]
          state?: string
          status?: Database["public"]["Enums"]["user_status_enum"]
          street?: string
          zipcode?: string
        }
        Relationships: [
          {
            foreignKeyName: "users_state_fkey"
            columns: ["state"]
            isOneToOne: false
            referencedRelation: "states"
            referencedColumns: ["code"]
          },
        ]
      }
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never